apps/api/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg==3.2.6
django-environ==0.12.0
djangorestframework-simplejwt==5.5.0
